+++ src/components/BaseComps/BaseComps.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
 export const PageTitle = styled.h2`
-  color: ${p => p.theme.colors.textPrimary};
+  color: ${(p: { theme: { colors: { textPrimary: string; }; }; }) => p.theme.colors.textPrimary};
   text-align: center;
 `;
 
@@ -9,8 +9,8 @@ export const List = styled.ul`
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
-  margin-top: ${p => p.theme.space[4]}px;
-  gap: ${p => p.theme.space[3]}px;
+  margin-top: ${(p: { theme: { space: number[]; }; }) => p.theme.space[4]}px;
+  gap: ${(p: { theme: { space: number[]; }; }) => p.theme.space[3]}px;
   list-style: none;
 `;
 
@@ -22,6 +22,6 @@ export const Item = styled.li`
 
   :hover,
   :focus {
-    box-shadow: ${p => p.theme.shadows.blackShadow};
+    box-shadow: ${(p: { theme: { shadows: { blackShadow: string; }; }; }) => p.theme.shadows.blackShadow};
   }
 `;
