--- src/components/DetailsComps/DetailsComps.jsx
@@ -1,42 +1,0 @@
-import styled from '@emotion/styled';
-
-export const DetailsName = styled.h1`
-  margin-top: ${p => p.theme.space[4]}px;
-  font-size: ${p => p.theme.fontSizes.l};
-  color: ${p => p.theme.colors.textTitle};
-  text-align: center;
-`;
-
-export const DetailsImg = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  cursor: pointer;
-`;
-
-export const DetailsDescr = styled.p`
-  margin-top: ${p => p.theme.space[4]}px;
-  color: ${p => p.theme.colors.textPrimary};
-  
-
-  @media (max-width: 479.98px) {
-    text-align: center;
-  }
-`;
-
-export const Bold = styled.span`
-  font-weight: ${p => p.bold ? '700' : '400'};
-`;
-
-export const DetailsHomePage = styled.a`
-  margin-top: ${p => p.theme.space[4]}px;
-  color: ${p => p.theme.colors.textPrimary};
-  text-decoration: none;
-
-  transition: color 250ms linear;
-
-  :hover,
-  :focus {
-    color: ${p => p.theme.colors.accent};
-  }
-`;
