--- src/components/Overview/Overview.styled.jsx
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const OverviewText = styled.p`
-  margin-top: ${p => p.theme.space[4]}px;
-  font-size: ${p => p.theme.fontSizes.s};
-  color: ${p => p.theme.colors.textPrimary};
-`;
